No api_lib dependence in aries
